requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.1,<4.0
django-environ>=0.11,<1.0
google-auth>=2.25,<3.0
python-jose[cryptography]>=3.3,<4.0
gunicorn>=21.0,<23.0
Pillow>=10.0,<11.0
django-cors-headers>=4.3,<5.0
requests>=2.31,<3.0
whitenoise>=6.6,<7.0
